+++ iocontrol/pagination.go
@@ -1,0 +1,49 @@
+package iocontrol
+
+import (
+	"errors"
+)
+
+/*
+ * @param manLists オプションとオプション説明が格納された文字列と、各オプション説明の行数の配列
+ * @description 各ページの先頭となるオプション配列manListsのindex番号が格納された配列を生成する
+ */
+func (iocontroller *IoController) LocatePages(manLists []ManData) []int {
+	var maxLineNumber = -1
+	pageList := []int{0}
+	// query行、option stack行、SEPARATORの３行
+	var lineCount = 3
+	var page = 0
+	iocontroller.maxPage = 0
+
+	for i := 0; i < len(manLists); i++ {
+		// for文を抜けた後に、ウィンドウの高さが低すぎて描画できないかを判定するために、
+		// 一番行数の多いオプション説明文の行数を求める
+		if maxLineNumber < manLists[i].LineNumber {
+			maxLineNumber = manLists[i].LineNumber
+		}
+
+		// ウィンドウの高さをオーバーしてしまう場合、次のページにオプション説明を表示する
+		if iocontroller.height < lineCount+manLists[i].LineNumber {
+			lineCount = 2
+			page++
+			pageList = append(pageList, i)
+			if iocontroller.maxPage < page {
+				iocontroller.maxPage = page
+			}
+		}
+
+		lineCount += manLists[i].LineNumber
+
+		if i == len(manLists)-1 {
+			pageList = append(pageList, i+1)
+		}
+	}
+
+	// 2は、query行と先頭SEPARATORの2行分
+	if iocontroller.height < maxLineNumber+2 {
+		panic(errors.New("Window height is too small"))
+	}
+
+	return pageList
+}
